requirements.txt
qrcode
qrcode[pil]
reportlab
PyPDF2
pdf2image
pytesseract
pdfplumber
Pillow
